chore: use tRPC
