+++ electron/api.ts
@@ -1,8 +1,12 @@
 import z from 'zod';
 import { initTRPC } from '@trpc/server';
+import { observable } from '@trpc/server/observable';
+import { EventEmitter } from 'events';
 
-// 呼び出し元は service に集約したい
+// 呼び出し元は集約したい
 import * as service from './service';
+
+const ee = new EventEmitter();
 
 const t = initTRPC.create({ isServer: true });
 const { procedure } = t;
@@ -12,6 +16,19 @@ export const router = t.router({
   createTodo: procedure.input(z.object({ text: z.string() })).mutation(async (req) => {
     return req;
   }),
+  subscribeToast: t.procedure.subscription(() => {
+    return observable((emit) => {
+      function onToast(text: string) {
+        emit.next({ text });
+      }
+
+      ee.on('toast', onToast);
+
+      return () => {
+        ee.off('toast', onToast);
+      };
+    });
+  }),
   getVRChatLogFilesDir: procedure.query(async () => {
     const logFilesDir = service.getVRChatLogFilesDir();
     return logFilesDir;
@@ -19,6 +36,16 @@ export const router = t.router({
   getVRChatPhotoDir: procedure.query(async () => {
     const vrchatPhotoDir = service.getVRChatPhotoDir();
     return vrchatPhotoDir;
+  }),
+  getStatusToUseVRChatLogFilesDir: procedure.query(async () => {
+    const vrchatLogFilesDir = service.getVRChatLogFilesDir();
+    let status: 'ready' | 'logFilesDirNotSet' | 'logFilesNotFound' | 'logFileDirNotFound' = 'ready';
+    if (vrchatLogFilesDir.path === null) {
+      status = 'logFilesDirNotSet';
+    } else if (vrchatLogFilesDir.error !== null) {
+      status = vrchatLogFilesDir.error;
+    }
+    return status;
   })
 });
