+++ electron/index.ts
@@ -2,14 +2,12 @@
 import { join } from 'path';
 
 // Packages
-import { BrowserWindow, app, ipcMain, IpcMainEvent, dialog } from 'electron';
+import { BrowserWindow, app, ipcMain } from 'electron';
 import isDev from 'electron-is-dev';
-import { createIPCHandler } from 'electron-trpc/main'; // ←追加
-import { router } from './api'; // ←追加
+import { createIPCHandler } from 'electron-trpc/main';
+import { router } from './api';
 
-import * as settingStore from './settingStore';
-// 呼び出し元は service に集約したい
-import * as service from './service';
+import * as controller from './controller';
 
 const CHANNELS = {
   CLEAR_ALL_STORED_SETTINGS: 'clear-all-stored-settings',
@@ -28,163 +26,14 @@ const CHANNELS = {
   GET_STATUS_TO_USE_VRCHAT_PHOTO_DIR: 'get-status-to-use-vrchat-photo-dir'
 };
 
-const MESSAGE = {
-  STATUS_TO_USE_VRCHAT_LOG_FILES_DIR: 'status-to-use-vrchat-log-files-dir',
-  STATUS_TO_USE_VRCHAT_PHOTO_DIR: 'status-to-use-vrchat-photo-dir',
-  VRCHAT_PHOTO_DIR: 'vrchat-photo-dir',
-  VRCHAT_PHOTO_DIR_WITH_ERROR: 'vrchat-photo-dir-with-error'
-};
-
-const messages = {
-  PATH_NOT_SET: 'Path is not set',
-  LOG_PATH_SET: (path: string) => `Log file path set to ${path}`
-};
-
-const handleClearAllStoredSettings = () => {
-  settingStore.clearAllStoredSettings();
-};
-
-const handleOpenDialogAndSetLogFilesDir = (event: IpcMainEvent) => {
-  dialog
-    .showOpenDialog({
-      properties: ['openDirectory']
-    })
-    .then((result) => {
-      if (!result.canceled) {
-        const dirPath = result.filePaths[0];
-        settingStore.setLogFilesDir(dirPath);
-        event.sender.send(CHANNELS.TOAST, messages.LOG_PATH_SET(dirPath));
-        event.sender.send(CHANNELS.LOG_FILES_DIR, dirPath);
-      }
-    })
-    .catch((err) => {
-      console.log(err);
-      event.sender.send(CHANNELS.TOAST, err.message);
-    });
-};
-
-const handleGetLogFilesDir = (event: IpcMainEvent) => {
-  const logFilesDir = service.getVRChatLogFilesDir();
-  event.sender.send(CHANNELS.LOG_FILES_DIR, logFilesDir.storedPath);
-  event.sender.send(CHANNELS.LOG_FILES_DIR_WITH_ERROR, {
-    storedPath: logFilesDir.storedPath,
-    path: logFilesDir.path,
-    error: logFilesDir.error
-  });
-};
-
-const handlegetStatusToUseVRChatLogFilesDir = (event: IpcMainEvent) => {
-  const vrchatLogFilesDir = service.getVRChatLogFilesDir();
-  let status: 'ready' | 'logFilesDirNotSet' | 'logFilesNotFound' | 'logFileDirNotFound' = 'ready';
-  if (vrchatLogFilesDir.path === null) {
-    status = 'logFilesDirNotSet';
-  } else if (vrchatLogFilesDir.error !== null) {
-    status = vrchatLogFilesDir.error;
-  }
-  event.sender.send(MESSAGE.STATUS_TO_USE_VRCHAT_LOG_FILES_DIR, status);
-};
-
-const handleOpenDialogAndSetVRChatPhotoDir = (event: IpcMainEvent) => {
-  dialog
-    .showOpenDialog({
-      properties: ['openDirectory']
-    })
-    .then((result) => {
-      if (!result.canceled) {
-        const dirPath = result.filePaths[0];
-        settingStore.setVRChatPhotoDir(dirPath);
-        event.sender.send(MESSAGE.VRCHAT_PHOTO_DIR, dirPath);
-        event.sender.send(CHANNELS.TOAST, `VRChat photo path set to ${dirPath}`);
-      }
-    })
-    .catch((err) => {
-      console.log(err);
-    });
-};
-
-const handleGetVRChatPhotoDir = (event: IpcMainEvent) => {
-  const vrchatPhotoDir = service.getVRChatPhotoDir();
-  event.sender.send(MESSAGE.VRCHAT_PHOTO_DIR, vrchatPhotoDir.storedPath);
-  event.sender.send(MESSAGE.VRCHAT_PHOTO_DIR_WITH_ERROR, {
-    storedPath: vrchatPhotoDir.storedPath,
-    path: vrchatPhotoDir.path,
-    error: vrchatPhotoDir.error
-  });
-};
-
-const handleCreateFiles = (event: IpcMainEvent) => {
-  // get log lines
-  const logFilesDir = service.getVRChatLogFilesDir();
-  if (typeof logFilesDir.path !== 'string') {
-    event.sender.send('toast', `Log file path is not set`);
-    return;
-  }
-  if (logFilesDir.error !== null) {
-    switch (logFilesDir.error) {
-      case 'logFilesNotFound':
-        event.sender.send('toast', `Log files not found`);
-        break;
-      default:
-        event.sender.send('toast', `Unknown error: ${logFilesDir.error}`);
-        break;
-    }
-    return;
-  }
-  const convertWorldJoinLogInfoListResult = service.convertLogLinesToWorldJoinLogInfosByVRChatLogDir(logFilesDir.path);
-  if (convertWorldJoinLogInfoListResult.isErr()) {
-    switch (convertWorldJoinLogInfoListResult.error.code) {
-      case 'LOG_FILE_NOT_FOUND':
-        event.sender.send('toast', `Log file not found`);
-        break;
-      case 'LOG_FILE_DIR_NOT_FOUND':
-        event.sender.send('toast', `Log file dir not found`);
-        break;
-      case 'LOG_FILES_NOT_FOUND':
-        event.sender.send('toast', `Log files not found`);
-        break;
-      default:
-        event.sender.send('toast', `Unknown error: ${convertWorldJoinLogInfoListResult.error.code}`);
-        break;
-    }
-    return;
-  }
-  const convertWorldJoinLogInfoList = convertWorldJoinLogInfoListResult.value;
-
-  // create files
-  const vrchatPhotoDir = service.getVRChatPhotoDir();
-  if (typeof vrchatPhotoDir.storedPath !== 'string') {
-    event.sender.send('toast', `VRChat photo path is not set`);
-    return;
-  }
-  if (vrchatPhotoDir.error !== null) {
-    switch (vrchatPhotoDir.error) {
-      case 'photoYearMonthDirsNotFound':
-        event.sender.send('toast', `Photo year-month dirs not found`);
-        break;
-      default:
-        event.sender.send('toast', `Unknown error: ${vrchatPhotoDir.error}`);
-        break;
-    }
-    return;
-  }
-
-  try {
-    service.createFiles(vrchatPhotoDir.storedPath, convertWorldJoinLogInfoList);
-    event.sender.send('toast', `Files created`);
-  } catch (error) {
-    console.log(error);
-    event.sender.send('toast', `Error: ${error}`);
-  }
-};
-
 function registerIpcMainListeners() {
-  ipcMain.on(CHANNELS.OPEN_DIALOG_AND_SET_LOG_FILES_DIR, handleOpenDialogAndSetLogFilesDir);
-  ipcMain.on(CHANNELS.GET_LOG_FILES_DIR, handleGetLogFilesDir);
-  ipcMain.on(CHANNELS.OPEN_DIALOG_AND_SET_VRCHAT_PHOTO_DIR, handleOpenDialogAndSetVRChatPhotoDir);
-  ipcMain.on(CHANNELS.GET_VRCHAT_PHOTO_DIR, handleGetVRChatPhotoDir);
-  ipcMain.on(CHANNELS.CREATE_FILES, handleCreateFiles);
-  ipcMain.on(CHANNELS.GET_STATUS_TO_USE_VRCHAT_LOG_FILES_DIR, handlegetStatusToUseVRChatLogFilesDir);
-  ipcMain.on(CHANNELS.CLEAR_ALL_STORED_SETTINGS, handleClearAllStoredSettings);
+  ipcMain.on(CHANNELS.OPEN_DIALOG_AND_SET_LOG_FILES_DIR, controller.handleOpenDialogAndSetLogFilesDir);
+  ipcMain.on(CHANNELS.GET_LOG_FILES_DIR, controller.handleGetLogFilesDir);
+  ipcMain.on(CHANNELS.OPEN_DIALOG_AND_SET_VRCHAT_PHOTO_DIR, controller.handleOpenDialogAndSetVRChatPhotoDir);
+  ipcMain.on(CHANNELS.GET_VRCHAT_PHOTO_DIR, controller.handleGetVRChatPhotoDir);
+  ipcMain.on(CHANNELS.CREATE_FILES, controller.handleCreateFiles);
+  ipcMain.on(CHANNELS.GET_STATUS_TO_USE_VRCHAT_LOG_FILES_DIR, controller.handlegetStatusToUseVRChatLogFilesDir);
+  ipcMain.on(CHANNELS.CLEAR_ALL_STORED_SETTINGS, controller.handleClearAllStoredSettings);
   ipcMain.on(CHANNELS.MESSAGE, (_, message) => {
     console.log(message);
   });
