+++ src/page/MainContainer.tsx
@@ -1,6 +1,5 @@
-import React, { useEffect } from 'react';
-import toast from 'react-hot-toast';
-import { ArrowPathIcon } from '@heroicons/react/24/outline';
+import React from 'react';
+// import { ArrowPathIcon } from '@heroicons/react/24/outline';
 import { Link } from 'react-router-dom';
 import { trpcReact } from '../trpc';
 
@@ -8,21 +7,6 @@ function MainContainer() {
   // 初期表示時に log-files-dir を取得する
   const logFilesDir = trpcReact.getVRChatLogFilesDir.useQuery().data?.path;
   const vrchatPhotoDir = trpcReact.getVRChatPhotoDir.useQuery().data?.path;
-
-  useEffect(() => {
-    if (window.Main)
-      window.Main.on('file-content', (content: string) => {
-        console.log(content);
-      });
-  });
-
-  useEffect(() => {
-    if (window.Main)
-      window.Main.on('toast', (content: string) => {
-        console.log(content);
-        toast(content);
-      });
-  });
 
   return (
     <div className="flex-auto">
@@ -52,7 +36,7 @@ function MainContainer() {
         </button>
         <div className="vrchat-photo-dir-label">vrchat-photo-dir: {vrchatPhotoDir}</div>
         {/* 設定の再取得 */}
-        <button
+        {/* <button
           className="get-vrchat-photo-dir-button py-2 px-4 bg-white rounded focus:outline-none shadow hover:bg-yellow-200"
           onClick={() => {
             if (window.Main) {
@@ -62,7 +46,7 @@ function MainContainer() {
           }}
         >
           <ArrowPathIcon className="h-5 w-5" />
-        </button>
+        </button> */}
         {/* すべての設定をリセットする */}
         <Link to="/clear-settings">
           <button className="reset-button py-2 px-4 bg-white rounded focus:outline-none shadow hover:bg-yellow-200">
